renaming percentile to quantile
